requirements.txt
fairseq==0.12.2
transformers==4.35.2
accelerate==0.24.1
datasets[audio]==2.15.0
torch==2.1.1
torchaudio==2.1.1
XSLTools==0.5
huggingface-hub==0.19.4
scipy==1.11.4
pydantic==2.5.2
pydantic_core==2.14.5
langchain==0.0.342
langchain-core==0.0.7
openai==1.3.6
aiogram==2.25.2
aioschedule==0.5.2
python-telegram-bot==13.14
python-dotenv==0.21.0
g2p_en
